Publish endpoints of terminating pods
Before this commit, endpoints of terminating pods (that is, pods with
non-nil deletion timestamps) were published only if the associated service
specified the true value in its spec.publishNotReadyAddresses field.
However, spec.publishNotReadyAddresses controls whether the endpoint should
be published as "ready", not whether it should be published at all.

After this commit, endpoints for terminating pods are published in
endpoints under notReadyAddresses and in endpointslices with the false
value for the ready condition.

* pkg/controller/endpoint/endpoints_controller.go (syncService): Delete the
logic to check the pod's deletion timestamp.
* pkg/controller/endpointslice/reconciler.go (reconcile): Delete the
publishNotReady argument to ShouldPodBeInEndpoints.
* pkg/controller/util/endpoint/controller_utils.go
(ShouldPodBeInEndpoints): Delete the publishNotReady parameter and
associated logic.
* pkg/controller/util/endpoint/controller_utils_test.go
(TestShouldPodBeInEndpoints): Delete the publishNotReady test input and the
newly redundant test case.